Applying Nick Lamb's patch:
"Attached are two files, first is a replacement ladspa.h which adds all the
stuff from LADSPA 1.1, you could have updated this file from the latest
LADSPA SDK, but now there's no need.

Second is a patch for LadspaEffect.cpp which adds support for these great
new hints. Now the LADSPA controls will have default settings which are
appropriate for the individual plugin, so users spend less time trying to
find a starting point for experimenting with the controls."