Fix:
P2 	(was P3) Split New: If selecting part of a clip from the left edge of a clip starting at time zero, the newly split clip wrongly aligns with the left edge of the residual clip in the original track. (Different bug to above, this one applies to single as well as multiple tracks).