Modifications to script plugin:
Include a basic makefile based on one of Leland's. To be able to build the
plugin you need to fill in the Audacity path.

Linux - now there are separate pipes for each direction, same as the Windows one.
Incidentally the problem with the second read failing was due to the test script.

Temporarily disabled support for getting responses back from Audacity, because this isn't yet handled on the Audacity side.