This update provides improved structure for the commands system and adds a command for taking screenshots. The screenshot tools dialog is modified to use the command, and the command is made available to scripts.
The commands system now has a system of parameters which allows type-checking
and validation of data passed in by a script. This is necessary for more useful
error messages to be returned than can be provided by the batch chain system.

Each command class has a 'signature' consisting of a map from parameter name to
a 'validator' and a default value. The validator is an object which is capable
of checking that a provided value is acceptable.

Commands also now can use abstract 'output targets' which mean progress
information and status messages can be dealt with in different ways without the
commands having to be directly concerned with e.g. displaying a progress
dialog.

The screenshot command takes three parameters: CaptureMode, Background, and
FilePath. See the test script for usage examples.  The behaviour of the command
should be close to how the dialog worked - and the dialog *should* work
approximately the same as it did before. Those familiar with the previous
behaviour of the dialog might want to test it and report any regressions to me.

One thing I did change was to comment out the contents of the OnUIUpdate method
- it caused everything to slow to a halt on my machine whenever the screenshot
tools dialog was open. There's probably a good reason for its being there, so
let me know if it should go back in.