Fix for http://bugzilla.audacityteam.org/show_bug.cgi?id=630 based on investigation by Gabriel Pato.
This is quite an involved piece of code, and to check it is OK it should be checked with mixtures of stereo and mono, looped and non looped, selected region, no selected region.  (I did, but no harm trying variations on that).  Real time effects, if enabled, could also potentially interact with this, and are subject to change in this area of code.