Possible fix for bug #435
I say possible because I can't fully test it as my motherboard
audio device doesn't show up in Windows (don't know why yet).

So, because of that and because this "fix" needs a little discussion
amongst the troops, I've ifdef'd it with EXPERIMENTAL_HAVE_DEVICE_CHANGE
and have disabled it by default.

What is does is it sets up a device change listener and performs an
automatic rescan when a change is detected.  (That's the part that
needs discussion.)