Fix for non-responsive timeline and meters after a timer record
The problem is that the fix for bug #334 is a bit over zealous and
blocks ALL application level events instead of just the keyboard events.
This was blocking the new AUDIOIO event from being handled.

I didn't want to monkey with the #334 solution at this time, so the
simple fix was to ensure the progress dialog was going (thus, enabling
appilcation events) was destroyed before attempting to stop the audio
stream.